--- bot.py
@@ -1,218 +1,0 @@
-import gspread
-from google.oauth2.service_account import Credentials
-import os
-import random
-from aiogram import Bot, Dispatcher, types
-from aiogram.types import (
-    InlineKeyboardMarkup,
-    InlineKeyboardButton,
-    CallbackQuery,
-    Contact,
-)
-from aiogram.filters import Command
-import asyncio
-from datetime import datetime, timedelta
-from dotenv import load_dotenv
-
-load_dotenv()
-
-# Настройки бота
-API_TOKEN = os.getenv("API_TOKEN")
-CREDENTIALS_FILE = os.getenv("CREDENTIALS_FILE")
-ADMIN_IDS = [int(x) for x in os.getenv("ADMIN_IDS", "").split(',') if x.strip()]
-
-if not API_TOKEN:
-    raise EnvironmentError("API_TOKEN is not set")
-if not CREDENTIALS_FILE:
-    raise EnvironmentError("CREDENTIALS_FILE is not set")
-if not ADMIN_IDS:
-    raise EnvironmentError("ADMIN_IDS is not set")
-
-bot = Bot(token=API_TOKEN)
-dp = Dispatcher()
-
-RENT_COST_PER_SESSION = 330
-
-SCOPE = ["https://spreadsheets.google.com/feeds", "https://www.googleapis.com/auth/drive"]
-
-SPREADSHEET_NAME = "CRM_BOT"
-
-client = None
-sheet = None
-clients_sheet = None
-history_sheet = None
-groups_sheet = None
-
-def init_gspread() -> None:
-    """Initialize Google Sheets connection."""
-    global client, sheet, clients_sheet, history_sheet, groups_sheet
-    if not os.path.exists(CREDENTIALS_FILE):
-        raise FileNotFoundError(f"Файл {CREDENTIALS_FILE} не знайдено.")
-    creds = Credentials.from_service_account_file(CREDENTIALS_FILE, scopes=SCOPE)
-    client = gspread.authorize(creds)
-    sheet = client.open(SPREADSHEET_NAME)
-    clients_sheet = sheet.worksheet("Клієнти")
-    history_sheet = sheet.worksheet("История")
-    groups_sheet = sheet.worksheet("Группа")
-
-def get_client_name(row):
-    return row.get("Ім’я") or row.get("Імя") or row.get("Имя") or row.get("Имʼя") or "Клієнт"
-
-@dp.message(lambda message: message.contact)
-async def register_by_contact(message: types.Message):
-    user_id = str(message.contact.user_id)
-    name = message.contact.first_name or "Клієнт"
-    records = clients_sheet.get_all_records()
-    if any(str(row["ID"]) == user_id for row in records):
-        await message.answer("❗ Ви вже зареєстровані.")
-        return
-    end_date = (datetime.now() + timedelta(days=60)).strftime("%Y-%m-%d")
-    clients_sheet.append_row([user_id, name, 10, end_date, "-"])
-    print(f"[LOG] Зареєстровано нового користувача: {user_id} ({name})")
-    await message.answer("✅ Реєстрація успішна! Вам додано 10 занять на 60 днів.")
-
-@dp.message(Command(commands=["start"]))
-async def send_welcome(message: types.Message):
-    keyboard = [
-        [InlineKeyboardButton(text="📊 Мої заняття", callback_data="my_sessions")],
-        [InlineKeyboardButton(text="📜 Історія занять", callback_data="view_history")],
-        [InlineKeyboardButton(text="💳 Отримати абонемент", callback_data="request_subscription")],
-        [InlineKeyboardButton(text="✅ Відмітити заняття", callback_data="mark_session")],
-        [InlineKeyboardButton(text="😼 Таємна кнопка", callback_data="secret_button")]
-    ]
-    if message.from_user.id in ADMIN_IDS:
-        keyboard.insert(0, [InlineKeyboardButton(text="🔧 Панель адміністратора", callback_data="admin_panel")])
-    await message.answer("Вітаю в CRM боті 🐬", reply_markup=InlineKeyboardMarkup(inline_keyboard=keyboard))
-
-@dp.callback_query(lambda c: c.data == "my_sessions")
-async def my_sessions(callback: CallbackQuery):
-    user_id = str(callback.from_user.id)
-    records = clients_sheet.get_all_records()
-    for row in records:
-        if str(row["ID"]) == user_id:
-            print(f"[LOG] Перевірка занять для {user_id}")
-            await callback.message.answer(f"У вас залишилось {row['К-сть тренувань']} занять. Термін дії: {row['Кінцева дата']}")
-            return
-    await callback.message.answer("❗ Ви ще не зареєстровані.")
-
-@dp.callback_query(lambda c: c.data == "view_history")
-async def view_history(callback: CallbackQuery):
-    user_id = str(callback.from_user.id)
-    rows = history_sheet.get_all_values()
-    lines = [f"{row[1]} – {row[2]}" for row in rows if row[0] == user_id]
-    await callback.message.answer("\n".join(lines) if lines else "Історія пуста.")
-
-@dp.callback_query(lambda c: c.data == "mark_session")
-async def mark_session(callback: CallbackQuery):
-    keyboard = InlineKeyboardMarkup(inline_keyboard=[
-        [InlineKeyboardButton(text="✅ Підтвердити", callback_data=f"approve_deduction:{callback.from_user.id}"),
-         InlineKeyboardButton(text="❌ Скасувати", callback_data="cancel_request")]
-    ])
-    print(f"[LOG] Отримано запит на списання заняття від {callback.from_user.id}")
-    await bot.send_message(ADMIN_IDS[0], f"Запит на списання заняття від {callback.from_user.id}", reply_markup=keyboard)
-    await callback.message.answer("✅ Запит надіслано адміну")
-
-@dp.callback_query(lambda c: c.data == "cancel_request")
-async def cancel_request(callback: CallbackQuery):
-    await callback.message.edit_text("Запит скасовано.")
-
-@dp.callback_query(lambda c: c.data.startswith("approve_deduction:"))
-async def approve_deduction(callback: CallbackQuery):
-    user_id = callback.data.split(":")[1]
-    records = clients_sheet.get_all_records()
-    for idx, row in enumerate(records):
-        if str(row["ID"]) == user_id:
-            new_sessions = max(0, int(row["К-сть тренувань"]) - 1)
-            clients_sheet.update_cell(idx+2, 3, new_sessions)
-            history_sheet.append_row([user_id, datetime.now().strftime("%Y-%m-%d %H:%M"), "Списано 1 заняття"])
-            name = get_client_name(row)
-            print(f"[LOG] Списано 1 заняття для {user_id}, залишок: {new_sessions}")
-            await bot.send_message(user_id, f"❗ Списано 1 заняття. Залишок: {new_sessions}")
-            await callback.message.answer(f"Списано для {name} (ID: {user_id})")
-            return
-    await callback.message.answer("Клієнт не знайдений")
-
-@dp.callback_query(lambda c: c.data == "request_subscription")
-async def request_subscription(callback: CallbackQuery):
-    keyboard = InlineKeyboardMarkup(inline_keyboard=[
-        [InlineKeyboardButton(text="✅ Дати 10 занять", callback_data=f"approve_subscription:{callback.from_user.id}"),
-         InlineKeyboardButton(text="❌ Відмовити", callback_data="deny_subscription")]
-    ])
-    print(f"[LOG] Клієнт {callback.from_user.id} запитав абонемент")
-    await bot.send_message(ADMIN_IDS[0], f"Запит на новий абонемент від {callback.from_user.id}", reply_markup=keyboard)
-    await callback.message.answer("💳 Запит надіслано адміну")
-
-@dp.callback_query(lambda c: c.data == "deny_subscription")
-async def deny_subscription(callback: CallbackQuery):
-    await callback.message.edit_text("❌ Запит на абонемент відхилено.")
-
-@dp.callback_query(lambda c: c.data.startswith("approve_subscription:"))
-async def approve_subscription(callback: CallbackQuery):
-    user_id = callback.data.split(":")[1]
-    records = clients_sheet.get_all_records()
-    for idx, row in enumerate(records):
-        if str(row["ID"]) == user_id:
-            clients_sheet.update_cell(idx+2, 3, 10)
-            new_date = (datetime.now() + timedelta(days=60)).strftime("%Y-%m-%d")
-            clients_sheet.update_cell(idx+2, 4, new_date)
-            history_sheet.append_row([user_id, datetime.now().strftime("%Y-%m-%d %H:%M"), "Додано 10 занять"])
-            print(f"[LOG] Додано 10 занять користувачу {user_id}")
-            await bot.send_message(user_id, "🎉 Вам видано новий абонемент на 10 занять (60 днів)")
-            await callback.message.answer(f"Видано новий абонемент для ID {user_id}")
-            return
-    await callback.message.answer("Клієнт не знайдений")
-
-@dp.callback_query(lambda c: c.data == "admin_panel")
-async def admin_panel(callback: CallbackQuery):
-    clients = clients_sheet.get_all_records()
-    reserve_total = 0
-    for client in clients:
-        user_id = client["ID"]
-        name = get_client_name(client)
-        sessions = client["К-сть тренувань"]
-        reserve_total += sessions * RENT_COST_PER_SESSION
-        keyboard = InlineKeyboardMarkup(inline_keyboard=[
-            [InlineKeyboardButton(text="➕ Додати", callback_data=f"add_session:{user_id}"),
-             InlineKeyboardButton(text="➖ Списати", callback_data=f"approve_deduction:{user_id}")],
-            [InlineKeyboardButton(text="📜 Історія", callback_data=f"history:{user_id}")]
-        ])
-        await callback.message.answer(f"👤 {name} (ID: {user_id})\nЗалишок: {sessions}", reply_markup=keyboard)
-    await callback.message.answer(f"💰 Сума, яку потрібно тримати для оренди: {reserve_total} грн")
-
-@dp.callback_query(lambda c: c.data.startswith("add_session:"))
-async def add_session(callback: CallbackQuery):
-    user_id = callback.data.split(":")[1]
-    records = clients_sheet.get_all_records()
-    for idx, row in enumerate(records):
-        if str(row["ID"]) == user_id:
-            new_sessions = int(row["К-сть тренувань"]) + 1
-            clients_sheet.update_cell(idx+2, 3, new_sessions)
-            if not row.get("Ім’я"):
-                clients_sheet.update_cell(idx+2, 2, "Клієнт")
-            history_sheet.append_row([user_id, datetime.now().strftime("%Y-%m-%d %H:%M"), "Додано 1 заняття"])
-            await bot.send_message(user_id, f"➕ Вам додано 1 заняття. Тепер у вас {new_sessions}")
-            await callback.message.answer("Заняття додано")
-            return
-
-@dp.callback_query(lambda c: c.data.startswith("history:"))
-async def user_history(callback: CallbackQuery):
-    user_id = callback.data.split(":")[1]
-    rows = history_sheet.get_all_values()
-    lines = [f"{row[1]} – {row[2]}" for row in rows if row[0] == user_id]
-    await callback.message.answer("\n".join(lines) if lines else "Історія пуста.")
-
-@dp.callback_query(lambda c: c.data == "secret_button")
-async def secret_button(callback: CallbackQuery):
-    messages = [
-        "😾 Котик незадоволений, що ти натиснув цю кнопку!",
-        "🌀 Ти відкрив портал у котячий вимір... але нічого не сталося.",
-        "💥 А ти справді думав, що тут щось є?",
-        "🐾 Секретів тут нема, тільки хвости."
-    ]
-    await callback.message.answer(random.choice(messages))
-
-if __name__ == "__main__":
-    init_gspread()
-    print("[LOG] Бот запущено")
-    import asyncio
-    asyncio.run(dp.start_polling(bot))
